fetch nft data on loading screen to avoid page delay
