+++ client/src/pages/Page.js
@@ -1,32 +1,9 @@
-import axios from 'axios';
-import React from 'react';
-import { useState, useEffect } from 'react';
-
-const contract_address_temp = '0x6fa9f4b50e2950a8137a76649193816fb29dad2c';
-const token_id_temp = '7981';
+import React, { useState } from 'react';
+import { useLocation } from 'react-router-dom';
 
 const Page = () => {
-  const [details, setDetails] = useState([]);
-
-  // on page load
-  useEffect(() => {
-    // use query paramters her
-    getDetails();
-  }, []);
-
-  // use async fn to request
-  const getDetails = async () => {
-    // this will need to get passed by the NFC metadata
-    const contract_address = contract_address_temp;
-    const token_id = token_id_temp;
-
-    // build request url
-    const url = `/details/${contract_address}/${token_id}`;
-
-    // get request from express backend
-    const res = await axios.get(url);
-    setDetails(res.data);
-  };
+  const { state } = useLocation();
+  const [details, setDetails] = useState(state);
 
   return (
     <>
